Random Fourier feature basis

Generates a random Fourier feature basis matrix for a provided kernel, optionally rescaling the data to lie in the unit hypercube. Random features are of the form $$ \phi(x) = \cos(\omega^T x + b), $$ where \(\omega\) is a vector of frequencies sampled from the Fourier transform of the kernel, and \(b\sim\mathrm{Unif}[-\pi, \pi]\) is a random phase shift. The input data x may be shifted and rescaled before the feature mapping is applied, according to the stdize argument.

Usage

b_rff(
  ...,
  p = 100,
  kernel = k_rbf(),
  stdize = c("scale", "box", "symbox", "none"),
  n_approx = nextn(4 * p),
  freqs = NULL,
  phases = NULL,
  shift = NULL,
  scale = NULL
)

Arguments

...

The variable(s) to build features for. A single data frame or matrix may be provided as well.

p

The number of random features.

kernel

A kernel function. If one of the recognized kernel functions such as k_rbf() is provided, then the computations will be exact. Otherwise, the fast Fourier transform of the provided kernel function is used to generate the random features. The kernel should be shift-invariant and decay to zero at positive and negative infinity.

stdize

How to standardize the predictors, if at all. The default "scale" applies scale() to the input so that the features have mean zero and unit variance, "box" scales the data along each dimension to lie in the unit hypercube, and "symbox" scales the data along each dimension to lie in \([-0.5, 0.5]^d\).

n_approx

The number of discrete frequencies to use in calculating the Fourier transform of the provided kernel. Not used for certain kernels for which an analytic Fourier transform is available; see above.

freqs

Matrix of frequencies to use, ncol(freqs) matching the number of predictors. If provided, overrides those calculated automatically, thus ignoring p and kernel.

phases

Vector of phase shifts to use. If provided, overrides those calculated automatically, thus ignoring p and kernel.

shift

Vector of shifts, or single shift value, to use. If provided, overrides those calculated according to stdize.

scale

Vector of scales, or single scale value, to use. If provided, overrides those calculated according to stdize.

Value

A matrix of random Fourier features.

References

Rahimi, A., & Recht, B. (2007). Random features for large-scale kernel machines. Advances in neural information processing systems, 20.

Examples

data(quakes)

lm(depth ~ b_rff(lat, long, p=16), quakes)
#> 
#> Call:
#> lm(formula = depth ~ b_rff(lat, long, p = 16), data = quakes)
#> 
#> Coefficients:
#>                (Intercept)   b_rff(lat, long, p = 16)1  
#>                    1776.24                     2076.33  
#>  b_rff(lat, long, p = 16)2   b_rff(lat, long, p = 16)3  
#>                     964.69                      -39.91  
#>  b_rff(lat, long, p = 16)4   b_rff(lat, long, p = 16)5  
#>                    -387.45                    -2927.69  
#>  b_rff(lat, long, p = 16)6   b_rff(lat, long, p = 16)7  
#>                     236.78                      697.56  
#>  b_rff(lat, long, p = 16)8   b_rff(lat, long, p = 16)9  
#>                     197.52                     -129.22  
#> b_rff(lat, long, p = 16)10  b_rff(lat, long, p = 16)11  
#>                   -2066.33                     -377.20  
#> b_rff(lat, long, p = 16)12  b_rff(lat, long, p = 16)13  
#>                    3009.06                      155.98  
#> b_rff(lat, long, p = 16)14  b_rff(lat, long, p = 16)15  
#>                   -5329.97                      126.53  
#> b_rff(lat, long, p = 16)16  
#>                    -730.68  
#> 
lm(depth ~ b_rff(lat, long, p=8, kernel=k_rbf(scale=0.5)), quakes)
#> 
#> Call:
#> lm(formula = depth ~ b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5)), 
#>     data = quakes)
#> 
#> Coefficients:
#>                                           (Intercept)  
#>                                               212.478  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))1  
#>                                                -5.066  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))2  
#>                                               417.081  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))3  
#>                                                38.154  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))4  
#>                                              -103.652  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))5  
#>                                              -277.305  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))6  
#>                                              -330.216  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))7  
#>                                               -16.102  
#> b_rff(lat, long, p = 8, kernel = k_rbf(scale = 0.5))8  
#>                                              -438.076  
#> 
lm(depth ~ b_rff(lat, long, p=20, kernel=k_rbf(scale=5), stdize="none"), quakes)
#> 
#> Call:
#> lm(formula = depth ~ b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), 
#>     stdize = "none"), data = quakes)
#> 
#> Coefficients:
#>                                                            (Intercept)  
#>                                                                 424.02  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")1  
#>                                                               -1928.43  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")2  
#>                                                                 614.98  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")3  
#>                                                                 637.60  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")4  
#>                                                                -879.01  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")5  
#>                                                              -18519.07  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")6  
#>                                                                 -73.48  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")7  
#>                                                               -2701.68  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")8  
#>                                                                2315.78  
#>  b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")9  
#>                                                                 -77.36  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")10  
#>                                                               -8343.52  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")11  
#>                                                                 193.50  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")12  
#>                                                                2735.79  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")13  
#>                                                                -195.64  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")14  
#>                                                               -3171.10  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")15  
#>                                                               -2196.33  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")16  
#>                                                                4073.88  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")17  
#>                                                                 328.49  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")18  
#>                                                                  81.56  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")19  
#>                                                               -7651.58  
#> b_rff(lat, long, p = 20, kernel = k_rbf(scale = 5), stdize = "none")20  
#>                                                                -425.00  
#>